backend/requirements.txt
fastapi
uvicorn
pydantic-settings
groq
langchain-groq
httpx
langgraph
python-multipart
tavily-python
